Added quiz structure to main
