+++ src/pages/Quiz/quiz.css
@@ -1,0 +1,86 @@
+:root {
+  --color-bg: #f9faff;
+  --color-card: #ffffff;
+  --color-primary: #5c6ac4;
+  --color-primary-hover: #4a54a1;
+  --color-text: #333333;
+  --color-muted: #777777;
+  --radius: 8px;
+  --transition: 0.3s ease-in-out;
+}
+
+body {
+  background: var(--color-bg);
+  font-family: 'Inter', sans-serif;
+  color: var(--color-text);
+}
+
+.question-block {
+  background: var(--color-card);
+  border: none;
+  border-radius: var(--radius);
+  padding: 1.5rem;
+  box-shadow: 0 4px 12px rgba(0, 0, 0, 0.05);
+  opacity: 0;
+  transform: translateY(20px);
+  animation: fadeInUp var(--transition) forwards;
+}
+
+@keyframes fadeInUp {
+  to {
+    opacity: 1;
+    transform: translateY(0);
+  }
+}
+
+.next-button {
+  background: var(--color-primary);
+  color: white;
+  border: none;
+  border-radius: var(--radius);
+  padding: 0.75rem 1.5rem;
+  font-size: 1rem;
+  cursor: pointer;
+  box-shadow: 0 2px 6px rgba(0, 0, 0, 0.1);
+  transition: transform var(--transition), background var(--transition);
+}
+
+.next-button:hover {
+  background: var(--color-primary-hover);
+  transform: scale(1.05);
+}
+
+
+.quiz-container {
+  max-width: 600px;
+  margin: 0 auto;
+  font-family: sans-serif;
+  padding: 1rem;
+}
+
+.question-block {
+  margin-bottom: 1.5rem;
+  padding: 0.5rem;
+  border: 1px solid #ddd;
+  border-radius: 4px;
+}
+
+.option-label {
+  display: block;
+  margin: 0.3rem 0;
+  cursor: pointer;
+}
+
+.submit-button {
+  padding: 0.5rem 1rem;
+  font-size: 1rem;
+  cursor: pointer;
+}
+
+.result-container {
+  max-width: 600px;
+  margin: 0 auto;
+  font-family: sans-serif;
+  padding: 2rem;
+  text-align: center;
+}
